requirements.txt
# Expense Control Backend — FastAPI + SQLAlchemy 2.0
# Python 3.11+
#
# Instalar: pip install -r requirements.txt

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.115.8
uvicorn[standard]==0.34.0
starlette==0.45.3
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic-core==2.41.5

# ============================================================================
# Database & ORM
# ============================================================================
SQLAlchemy==2.0.48
sqlmodel==0.0.22
sqladmin==0.20.1
alembic==1.18.4
asyncpg==0.31.0
psycopg2-binary==2.9.11

# ============================================================================
# Authentication & Security
# ============================================================================
PyJWT==2.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
cryptography==46.0.5
email-validator==2.2.0

# ============================================================================
# Utilities & Helpers
# ============================================================================
python-dotenv==1.2.2
python-multipart==0.0.20
httpx==0.28.1
requests==2.32.5

# ============================================================================
# Development & Testing
# ============================================================================
pytest==9.0.2
pytest-asyncio==1.3.0

# ============================================================================
# Optional: Error Tracking & Monitoring
# ============================================================================
sentry-sdk==2.52.0

# ============================================================================
# Transitive Dependencies (included automatically, listed for clarity)
# ============================================================================
# annotated-types, anyio, attrs, certifi, cffi, charset-normalizer,
# click, colorama, h11, httpcore, httptools, idna, inflection,
# iniconfig, jsonschema, packaging, pluggy, pyasn1, pycparser,
# Pygments, typing-extensions, tzdata, urllib3, websockets, etc.
